+++ types/part_set.go
@@ -21,7 +21,7 @@ var (
 )
 
 type Part struct {
-	Index int64              `json:"index"`
+	Index int32              `json:"index"`
 	Bytes tmbytes.HexBytes   `json:"bytes"`
 	Proof merkle.SimpleProof `json:"proof"`
 }
@@ -83,13 +83,13 @@ func (psh PartSetHeader) ValidateBasic() error {
 //-------------------------------------
 
 type PartSet struct {
-	total int64
+	total int32
 	hash  []byte
 
 	mtx           sync.Mutex
 	parts         []*Part
 	partsBitArray *bits.BitArray
-	count         int64
+	count         int32
 }
 
 // Returns an immutable, full PartSet from the data bytes.
@@ -102,7 +102,7 @@ func NewPartSetFromData(data []byte, partSize int) *PartSet {
 	partsBitArray := bits.NewBitArray(total)
 	for i := 0; i < total; i++ {
 		part := &Part{
-			Index: int64(i),
+			Index: int32(i),
 			Bytes: data[i*partSize : tmmath.MinInt(len(data), (i+1)*partSize)],
 		}
 		parts[i] = part
@@ -115,11 +115,11 @@ func NewPartSetFromData(data []byte, partSize int) *PartSet {
 		parts[i].Proof = *proofs[i]
 	}
 	return &PartSet{
-		total:         int64(total),
+		total:         int32(total),
 		hash:          root,
 		parts:         parts,
 		partsBitArray: partsBitArray,
-		count:         int64(total),
+		count:         int32(total),
 	}
 }
 
@@ -171,14 +171,14 @@ func (ps *PartSet) HashesTo(hash []byte) bool {
 	return bytes.Equal(ps.hash, hash)
 }
 
-func (ps *PartSet) Count() int64 {
+func (ps *PartSet) Count() int32 {
 	if ps == nil {
 		return 0
 	}
 	return ps.count
 }
 
-func (ps *PartSet) Total() int64 {
+func (ps *PartSet) Total() int32 {
 	if ps == nil {
 		return 0
 	}
